+++ src/Livewire/Chat/Group/AddMembers.php
@@ -9,7 +9,7 @@ use Livewire\Attributes\Validate;
 use Livewire\WithFileUploads;
 use Namu\WireChat\Facades\WireChat;
 use Namu\WireChat\Livewire\Chat\Info;
-use Namu\WireChat\Livewire\Modals\ModalComponent;
+use Namu\WireChat\Livewire\Concerns\ModalComponent;
 use Namu\WireChat\Models\Conversation;
 use Namu\WireChat\Models\Participant;
