+++ src/Livewire/Chat/Group/Members.php
@@ -11,7 +11,7 @@ use Livewire\WithPagination;
 use Namu\WireChat\Enums\Actions;
 use Namu\WireChat\Enums\ParticipantRole;
 use Namu\WireChat\Facades\WireChat;
-use Namu\WireChat\Livewire\Modals\ModalComponent;
+use Namu\WireChat\Livewire\Concerns\ModalComponent;
 use Namu\WireChat\Livewire\WireChat as WidgetsWireChat;
 use Namu\WireChat\Livewire\Chat\Info;
 use Namu\WireChat\Models\Action;
